I found an error with the code linking the HFSS regulation category file from Kantar with the master data file.
The revised code starts on row 285 to row 301 and creates two Kantar regulation category files: th first based on those rows which have a value in the extended_desc column and the second based on rows which had only NA values in the extended_desc column.

This allowed me to merge the master category file (market_sub_extend) with the two Kantar reg files separately: the first based on 3 IDS variables and the second based on 2 ID variables. I then appended them together.

The final stage was to merge this with the master file so that all HFSS purchases were correctly identified as being under the regulations or not