SIMD -postcode look up for Kantar updated to remove postcodes with B or C suffixes. Those with A suffixes were kept as this is where the population is weighted. The A suffix was then removed and the postcode concatenated back into one variable.
There are still 346 live postcodes without SIMD information. The SIMD - postcode lookup available appears to be out of date and there is a notice that it is due to be updated this month. The final file for Kantar would need to be refereshed with this new information before sending to Kantar.